+++ run-py-tests.sh
@@ -5,4 +5,5 @@ python3 -m venv venv
 . ./activate
 pip3 install .
 
-py.test ./tests -s -v
+py.test ./tests/core -s -v
+py.test ./tests/wallet -s -v
